+++ RELEASE_NOTES.md
@@ -14,6 +14,8 @@
 * Trims `StringValue` when coercing to primitive types
 * Fixes `Evaluator.finalizeInputValue` for compound and optional types
 * `Evaluator.finalizeInputValue` loads file contents from remote file source if file does not exist locally
+* `CwlType.coerceTo` now returns both the coerced-to type and value
+* Added `CwlType.CwlGenericRecord`, which is coercible to either `CwlInputRecord` or `CwlOutputRecord`
 
 ## 0.7.4 (2021-01-05)